+++ tests/ui/undocumented_unsafe_blocks.rs
@@ -472,19 +472,6 @@ mod unsafe_impl_invalid_comment {
     unsafe impl Interference for () {}
 }
 
-mod unsafe_items_invalid_comment {
-    // SAFETY:
-    const CONST: u32 = 0;
-    // SAFETY:
-    static STATIC: u32 = 0;
-    // SAFETY:
-    struct Struct;
-    // SAFETY:
-    enum Enum {}
-    // SAFETY:
-    mod module {}
-}
-
 unsafe trait ImplInFn {}
 
 fn impl_in_fn() {
@@ -520,37 +507,6 @@ fn issue_9142() {
             bar
         }
     };
-}
-
-mod unnecessary_from_macro {
-    trait T {}
-
-    macro_rules! no_safety_comment {
-        ($t:ty) => {
-            impl T for $t {}
-        };
-    }
-
-    // FIXME: This is not caught
-    // Safety: unnecessary
-    no_safety_comment!(());
-
-    macro_rules! with_safety_comment {
-        ($t:ty) => {
-            // Safety: unnecessary
-            impl T for $t {}
-        };
-    }
-
-    with_safety_comment!(i32);
-}
-
-fn unnecessary_on_stmt_and_expr() -> u32 {
-    // SAFETY: unnecessary
-    let num = 42;
-
-    // SAFETY: unnecessary
-    24
 }
 
 fn main() {}
